+++ src/common/input.ts
@@ -3482,31 +3482,34 @@ export class Input extends EventEmitter {
         if (frag == null) frag = false;
         this.buildTriggerCache();
         let t = 0;
-        const tl = this._TriggerCache.length;
+        //scope to get performance
+        const triggers = this._TriggerCache;
+        const tl = triggers.length;
         for (; t < tl; t++) {
-            if (this._TriggerCache[t].type !== undefined && this._TriggerCache[t].type !== type) continue;
-            if (frag && !this._TriggerCache[t].triggerPrompt) continue;
-            if (!frag && !this._TriggerCache[t].triggerNewline && (this._TriggerCache[t].triggerNewline !== undefined))
+            const trigger = triggers[t];
+            if (trigger.type !== undefined && trigger.type !== type) continue;
+            if (frag && !trigger.triggerPrompt) continue;
+            if (!frag && !trigger.triggerNewline && (trigger.triggerNewline !== undefined))
                 continue;
-            if (this._TriggerCache[t].verbatim) {
-                if (!this._TriggerCache[t].caseSensitive && raw.toLowerCase() !== this._TriggerCache[t].pattern.toLowerCase()) continue;
-                else if (this._TriggerCache[t].caseSensitive && raw !== this._TriggerCache[t].pattern) continue;
+            if (trigger.verbatim) {
+                if (!trigger.caseSensitive && raw.toLowerCase() !== trigger.pattern.toLowerCase()) continue;
+                else if (trigger.caseSensitive && raw !== trigger.pattern) continue;
                 if (ret)
-                    return this.ExecuteTrigger(this._TriggerCache[t], [raw], true, t);
-                this.ExecuteTrigger(this._TriggerCache[t], [raw], false, t);
+                    return this.ExecuteTrigger(trigger, [raw], true, t);
+                this.ExecuteTrigger(trigger, [raw], false, t);
             }
             else {
                 try {
                     let re;
-                    if (this._TriggerCache[t].caseSensitive)
-                        re = new RegExp(this._TriggerCache[t].pattern, 'g');
+                    if (trigger.caseSensitive)
+                        re = new RegExp(trigger.pattern, 'g');
                     else
-                        re = new RegExp(this._TriggerCache[t].pattern, 'gi');
+                        re = new RegExp(trigger.pattern, 'gi');
                     const res = re.exec(raw);
                     if (!res || !res.length) continue;
                     if (ret)
-                        return this.ExecuteTrigger(this._TriggerCache[t], res, true, t);
-                    this.ExecuteTrigger(this._TriggerCache[t], res, false, t);
+                        return this.ExecuteTrigger(trigger, res, true, t);
+                    this.ExecuteTrigger(trigger, res, false, t);
                 }
                 catch (e) {
                     if (this.client.options.showScriptErrors)