+++ changelog.md
@@ -12,6 +12,7 @@
   - Clear local cache every time the preferences are changed
   - Clear local cache when leaving combat if it has been at least an hour since last time
   - Display: tweak performance and memory usage to try and balance between not using to much memory and using enough to get performance gains
+  - Input: Performance tweaks to executing triggers
 - **Fixed:**
   - Code editor:
     - Area designer: